requirements/ci.txt
# SHA1:85131898e41a03f79cd3c530efa2bf85bb01b435
#
# This file is autogenerated by pip-compile-multi
# To update, run:
#
#    pip-compile-multi
#
distlib==0.3.9
    # via virtualenv
filelock==3.16.1
    # via
    #   tox
    #   virtualenv
packaging==24.1
    # via tox
platformdirs==4.3.6
    # via virtualenv
pluggy==1.5.0
    # via tox
py==1.11.0
    # via tox
ruamel-yaml==0.17.40
    # via tox-conda
ruamel-yaml-clib==0.2.8
    # via ruamel-yaml
six==1.16.0
    # via tox
tox==3.28.0
    # via
    #   -r requirements/ci.in
    #   tox-conda
tox-conda==0.10.2
    # via -r requirements/ci.in
virtualenv==20.26.6
    # via tox
